Commands that need i/ flag will prompt for it

+++ src/main/java/seedu/command/interpreter/EventCommandInterpreter.java
@@ -61,29 +61,44 @@ public class EventCommandInterpreter extends CommandInterpreter {
             command = new AddEvent(event, this.eventList);
             break;
         case "editname":
+            if (flagDoesNotExist(commandParameters, "i/")) {
+                throw new DukeException("EventCommandInterpreter: i/ flag is necessary");
+            }
             eventParser.parse(commandParameters);
             index = eventParser.getIndex();
             name = eventParser.getName();
             command = new EditName(index, name, this.eventList);
             break;
         case "editdatetime":
+            if (flagDoesNotExist(commandParameters, "i/")) {
+                throw new DukeException("EventCommandInterpreter: i/ flag is necessary");
+            }
             eventParser.parse(commandParameters);
             index = eventParser.getIndex();
             datetime = eventParser.getDate() + " " + eventParser.getTime();
             command = new EditDateTime(index, datetime, this.eventList);
             break;
         case "editvenue":
+            if (flagDoesNotExist(commandParameters, "i/")) {
+                throw new DukeException("EventCommandInterpreter: i/ flag is necessary");
+            }
             eventParser.parse(commandParameters);
             index = eventParser.getIndex();
             venue = eventParser.getVenue();
             command = new EditVenue(index, venue, this.eventList);
             break;
         case "editevent":
+            if (flagDoesNotExist(commandParameters, "i/")) {
+                throw new DukeException("EventCommandInterpreter: i/ flag is necessary");
+            }
             event = eventParser.parseEvent(commandParameters);
             index = eventParser.getIndex();
             command = new EditEvent(index, event, this.eventList);
             break;
         case "delete":
+            if (flagDoesNotExist(commandParameters, "i/")) {
+                throw new DukeException("EventCommandInterpreter: i/ flag is necessary");
+            }
             eventParser.parse(commandParameters);
             index = eventParser.getIndex();
             command = new DeleteEvent(index, this.eventList);
